NI_DAQ Coding
AO output seems to work fine on a different NI-DAQ unit, perhaps the one I was using was faulty. Can roughly get the NI-DAQ to perform a waveform output followed by a waveform read.

Will need to do some work on getting the timings to match between the writing and the reading.

In any case it should be possible to generate PRBS as you would like in order to be able to calibrate the pulse counting for the PMac Statistics experiment.